requirements.lock.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
anyio==4.13.0
async-timeout==5.0.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.7
click==8.3.2
clip @ git+https://github.com/openai/CLIP.git@d05afc436d78f1c48dc0dbf8e5980a9d471f35f6
datasets==4.8.4
diffusers==0.37.1
dill==0.4.1
dreamsim==0.2.1
einops==0.8.2
exceptiongroup==1.3.1
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
ftfy==6.3.1
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.11.0
idna==3.11
ImageIO==2.37.3
importlib_metadata==9.0.0
Jinja2==3.1.6
lazy-loader==0.5
lpips==0.1.4
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
networkx==3.4.2
numpy==2.2.6
open_clip_torch==3.3.0
packaging==26.0
pandas==2.3.3
pillow==12.1.1
propcache==0.4.1
psutil==7.2.2
pyarrow==23.0.1
Pygments==2.20.0
python-dateutil==2.9.0.post0
pytz==2026.1.post1
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
rich==15.0.0
safetensors==0.7.0
scikit-image==0.25.2
scipy==1.15.3
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tifffile==2025.5.10
tokenizers==0.22.2
torch==2.11.0+cpu
torchaudio==2.11.0+cpu
torchvision==0.26.0+cpu
tqdm==4.67.3
transformers==5.5.4
typer==0.24.1
typing_extensions==4.15.0
tzdata==2026.1
urllib3==2.6.3
wcwidth==0.6.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.1
